requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
chainer==2.0.2
filelock==2.0.12          # via chainer
mock==2.0.0               # via chainer
nose==1.3.7               # via chainer
numpy==1.13.1             # via chainer
olefile==0.44             # via pillow
pbr==3.1.1                # via mock
pillow==4.2.1
protobuf==3.4.0           # via chainer
six==1.10.0               # via chainer, mock, protobuf
